new: document SeeClickFix, CivicPlus, and app store publisher info
This is a landing page for the privacy statement link on mobile
apps provided by SeeClickFix (a division of CivicPlus)
The page provides an explanation for the relationship between
the organization listed as the developer/publisher,
SeeClickFix, and CivicPlus.

The actual privacy policy is linked and resides on the CivicPlus
website.